Updated Source Path
Changed the source path for the essay files, so that it's now hard-coded
to look for them in "http://lairbob.com/essays-html/"

A better answer would obviously be to at least infer the name of the
server URL, and hard-code a descendant path from there. The _best_
answer would be to make the "master path" a setting, within the plugin's
UI admin panel in WP.

Someday.